Add `Expectation` class as parent for new Expectations
