+++ src/Corretto/Expectation.php
@@ -1,0 +1,8 @@
+<?php
+namespace Corretto;
+
+class Expectation {
+	function __construct( $actual ) {
+		$this->actual = $actual;
+	}
+}
